Check finished tween id before delete it
This commit fixes this issue:

**Steps to reproduce the issue:**
- Create Tween A (it's added to tweens array with index = 0)
- Create Tween B (it's added to tweens array with index = 1) with the same duration as A
- In Tween A onComplete event:
    * Cancel Tween B
    * Create Tween C (it's added to tweens array with index = 0)
    * Create Tween D (it's added to tweens array with index = 1)

I know it's a rare case, but as I'm using tweens in a lot of objects, I have found this situation.

**What happens?**
Tween D is cancelled immediately after it has been created.

**Why?**
In the frame when Tween A and Tween B finish, LeanTween's update method adds 0 and 1 to tweensFinished array. Then, it iterates over tweensFinished array deleting each tween, and executing its onComplete method. So, in the first iteration, Tween A is deleted, and then, Tween A onComplete method is executed, where we cancel Tween B. So, now, Tween A and B are both canceled. There are no active tweens, so, we create Tween C, and Tween D, adding them to tweens array with index 0 and 1. 
Now, the next iteration is executed (we are going to cancel tweensFinished[1]). But the tween at tweensFinished[1] is not the same tween it was when it as added to tweensFinished, and we are cancelling Tween D unintentionally.

**The fix**
I think an easy way to fix it is to check the tween id before cancel it, so we are sure we are deleting the tween we wanted.